Moved CreateTagCommand to own NS
