+++ src/tests/Equinor.Procosys.Preservation.Command.Tests/TagCommands/CreateTag/CreateTagCommandTests.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
+using Equinor.Procosys.Preservation.Command.TagCommands;
 using Equinor.Procosys.Preservation.Command.TagCommands.CreateTag;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
